fix: fix some trouble with values of Cluster ID. WARNING: see body
since numeric_limits in the current std library version is not constexpr, the code was failing. We use INT_MAX regardless of the passed type T as we will quite often use int.
Moreover, we set the first cluster ID to 0 for readibility, as our dimension will never overflow the numeric limit